Merge commit '879621a3304b7317e8197e64d0a8dc8c2acde8c1' into production/cultourseurope

+++ Vps/Util/TempCleaner.php
@@ -11,7 +11,7 @@ class Vps_Util_TempCleaner
         }
 
         foreach (new DirectoryIterator('application/temp') as $f) {
-            if ($f->isFile() && $f->getMTime() < time() - $cleanOlderThan) {
+            if ($f->isFile() && $f->getFilename() != '.gitignore' && $f->getMTime() < time() - $cleanOlderThan) {
                 unlink($f->getPathname());
             }
         }
